Added PaymentsDTO,mapper,and some minor changes

+++ server/src/main/java/com/cdac/orderease/repository/PaymentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cdac.orderease.entity.Payments;
 
-public interface PaymentRepository extends JpaRepository<Payments, Integer>{
+public interface PaymentRepository extends JpaRepository<Payments, Long>{
 
 	Payments findByRazorpayOrderId(String razorpayId);
